+++ app/src/main/java/io/github/droidkaigi/confsched/viewmodel/SessionDetailViewModel.java
@@ -1,14 +1,25 @@
 package io.github.droidkaigi.confsched.viewmodel;
 
 import android.content.Context;
+import android.databinding.BindingAdapter;
+import android.support.annotation.NonNull;
+import android.support.design.widget.FloatingActionButton;
+import android.support.v4.content.ContextCompat;
 import android.text.TextUtils;
+import android.text.util.Linkify;
 import android.view.View;
+import android.widget.TextView;
+
+import java.util.Date;
 
 import javax.inject.Inject;
 
+import io.github.droidkaigi.confsched.R;
 import io.github.droidkaigi.confsched.dao.SessionDao;
 import io.github.droidkaigi.confsched.model.Session;
 import io.github.droidkaigi.confsched.util.AlarmUtil;
+import io.github.droidkaigi.confsched.util.DataBindingAttributeUtil;
+import io.github.droidkaigi.confsched.util.DateUtil;
 import io.github.droidkaigi.confsched.util.PageNavigator;
 import io.github.droidkaigi.confsched.viewmodel.event.EventBus;
 import io.github.droidkaigi.confsched.viewmodel.event.SessionSelectedChangedEvent;
@@ -72,6 +83,41 @@ public class SessionDetailViewModel implements ViewModel {
         // TODO This is not smart way. I want to solve by using two way binding.
         eventBus.post(new SessionSelectedChangedEvent(session));
         AlarmUtil.handleSessionAlarm(context, session);
+    }
+
+    public String sessionTimeRange() {
+        Date displaySTime = session.getDisplaySTime(context);
+        Date displayETime = session.getDisplayETime(context);
+
+        return context.getString(R.string.session_time_range,
+                DateUtil.getLongFormatDate(displaySTime, context),
+                DateUtil.getHourMinute(displayETime),
+                DateUtil.getMinutes(displaySTime, displayETime));
+    }
+
+    public int slideIconVisibility() {
+        return session.hasSlide() ? View.VISIBLE : View.GONE;
+    }
+
+    public int dashVideoIconVisibility() {
+        return session.hasDashVideo() ? View.VISIBLE : View.GONE;
+    }
+
+
+    /* ================================================== *
+     *  BindingAdapters
+     * ================================================== */
+    @BindingAdapter("sessionDescription")
+    public static void setSessionDescription(TextView textView, @NonNull Session session) {
+        DataBindingAttributeUtil.setTextRtlConsidered(textView, session.description);
+        Linkify.addLinks(textView, Linkify.ALL);
+    }
+
+    @BindingAdapter("sessionFab")
+    public static void setSessionFab(FloatingActionButton fab, @NonNull Session session) {
+        fab.setRippleColor(ContextCompat.getColor(fab.getContext(), session.category.getPaleColorResId()));
+        // TODO Use Observable field
+        fab.setSelected(session.checked);
     }
 
 }
